Add rudimentary colour support to Motifs
NB: This only works as intended if your colours are unambiguous when
considering subgraphs - e.g. the case where you have only one node of a
certain colour, and all of the other nodes are one other colour, and you
have subgraphs > size 2.

The fix is quite simple - we just need to add a "key" to each subgraph
as it is transformed by nauty, consisting of unconnected nodes of every
colour.